Move util function out of ibcclient.ts
