+++ src/lib/ibcclient.ts
@@ -1,4 +1,4 @@
-import { toAscii, toUtf8 } from '@cosmjs/encoding';
+import { toAscii } from '@cosmjs/encoding';
 import {
   buildFeeTable,
   Coin,
@@ -12,7 +12,6 @@ import { OfflineSigner, Registry } from '@cosmjs/proto-signing';
 import {
   AuthExtension,
   BankExtension,
-  BroadcastTxFailure,
   BroadcastTxResponse,
   defaultRegistryTypes,
   isBroadcastTxFailure,
@@ -26,19 +25,15 @@ import {
 import {
   adaptor34,
   CommitResponse,
-  ReadonlyDateWithNanoseconds,
   Header as RpcHeader,
-  ValidatorPubkey as RpcPubKey,
   Client as TendermintClient,
 } from '@cosmjs/tendermint-rpc';
 import { arrayContentEquals, sleep } from '@cosmjs/utils';
 import Long from 'long';
 
-import { HashOp, LengthOp } from '../codec/confio/proofs';
 import { Any } from '../codec/google/protobuf/any';
-import { Timestamp } from '../codec/google/protobuf/timestamp';
 import { MsgTransfer } from '../codec/ibc/applications/transfer/v1/tx';
-import { Order, Packet, State } from '../codec/ibc/core/channel/v1/channel';
+import { Order, State } from '../codec/ibc/core/channel/v1/channel';
 import {
   MsgChannelOpenAck,
   MsgChannelOpenConfirm,
@@ -62,7 +57,6 @@ import {
   ConsensusState as TendermintConsensusState,
   Header as TendermintHeader,
 } from '../codec/ibc/lightclients/tendermint/v1/tendermint';
-import { PublicKey as ProtoPubKey } from '../codec/tendermint/crypto/keys';
 import {
   blockIDFlagFromJSON,
   Commit,
@@ -72,6 +66,15 @@ import {
 import { ValidatorSet } from '../codec/tendermint/types/validator';
 
 import { IbcExtension, setupIbcExtension } from './queries/ibc';
+import {
+  buildClientState,
+  buildConsensusState,
+  createBroadcastTxErrorMessage,
+  mapRpcPubKeyToProto,
+  timestampFromDateNanos,
+  toIntHeight,
+  toProtoHeight,
+} from './utils';
 
 /**** These are needed to bootstrap the endpoints */
 /* Some of them are hardcoded various places, which should we make configurable? */
@@ -110,24 +113,6 @@ function ibcRegistry(): Registry {
     ['/ibc.core.channel.v1.MsgChannelOpenConfirm', MsgChannelOpenConfirm],
     ['/ibc.applications.transfer.v1.MsgTransfer', MsgTransfer],
   ]);
-}
-
-function timestampFromDateNanos(date: ReadonlyDateWithNanoseconds): Timestamp {
-  const nanos = (date.getTime() % 1000) * 1000000 + (date.nanoseconds ?? 0);
-  return Timestamp.fromPartial({
-    seconds: new Long(date.getTime() / 1000),
-    nanos,
-  });
-}
-
-export function toIntHeight(height?: Height): number {
-  return height?.revisionHeight?.toNumber() ?? 0;
-}
-
-export function toProtoHeight(height: number): Height {
-  return Height.fromPartial({
-    revisionHeight: new Long(height),
-  });
 }
 
 /// This is the default message result with no extra data
@@ -174,10 +159,6 @@ export interface ChannelHandshake {
 export interface ChannelInfo {
   readonly portId: string;
   readonly channelId: string;
-}
-
-function createBroadcastTxErrorMessage(result: BroadcastTxFailure): string {
-  return `Error when broadcasting tx ${result.transactionHash} at height ${result.height}. Code: ${result.code}; Raw log: ${result.rawLog}`;
 }
 
 export interface IbcFeeTable extends FeeTable {
@@ -941,25 +922,6 @@ export class IbcClient {
   }
 }
 
-function mapRpcPubKeyToProto(pubkey?: RpcPubKey): ProtoPubKey | undefined {
-  if (pubkey === undefined) {
-    return undefined;
-  }
-  if (pubkey.algorithm == 'ed25519') {
-    return {
-      ed25519: pubkey.data,
-      secp256k1: undefined,
-    };
-  } else if (pubkey.algorithm == 'secp256k1') {
-    return {
-      ed25519: undefined,
-      secp256k1: pubkey.data,
-    };
-  } else {
-    throw new Error(`Unknown validator pubkey type: ${pubkey.algorithm}`);
-  }
-}
-
 export interface CreateClientArgs {
   clientState: TendermintClientState;
   consensusState: TendermintConsensusState;
@@ -979,86 +941,6 @@ export async function buildCreateClientArgs(
     header.height
   );
   return { consensusState, clientState };
-}
-
-export function buildConsensusState(
-  header: RpcHeader
-): TendermintConsensusState {
-  return TendermintConsensusState.fromPartial({
-    timestamp: timestampFromDateNanos(header.time),
-    root: {
-      hash: header.appHash,
-    },
-    nextValidatorsHash: header.nextValidatorsHash,
-  });
-}
-
-// Note: we hardcode a number of assumptions, like trust level, clock drift, and assume revisionNumber is 1
-export function buildClientState(
-  chainId: string,
-  unbondingPeriodSec: number,
-  trustPeriodSec: number,
-  height: number
-): TendermintClientState {
-  // Copied here until https://github.com/confio/ics23/issues/36 is resolved
-  // https://github.com/confio/ics23/blob/master/js/src/proofs.ts#L11-L26
-  const iavlSpec = {
-    leafSpec: {
-      prefix: Uint8Array.from([0]),
-      hash: HashOp.SHA256,
-      prehashValue: HashOp.SHA256,
-      prehashKey: HashOp.NO_HASH,
-      length: LengthOp.VAR_PROTO,
-    },
-    innerSpec: {
-      childOrder: [0, 1],
-      minPrefixLength: 4,
-      maxPrefixLength: 12,
-      childSize: 33,
-      hash: HashOp.SHA256,
-    },
-  };
-  const tendermintSpec = {
-    leafSpec: {
-      prefix: Uint8Array.from([0]),
-      hash: HashOp.SHA256,
-      prehashValue: HashOp.SHA256,
-      prehashKey: HashOp.NO_HASH,
-      length: LengthOp.VAR_PROTO,
-    },
-    innerSpec: {
-      childOrder: [0, 1],
-      minPrefixLength: 1,
-      maxPrefixLength: 1,
-      childSize: 32,
-      hash: HashOp.SHA256,
-    },
-  };
-
-  return TendermintClientState.fromPartial({
-    chainId,
-    trustLevel: {
-      numerator: Long.fromInt(1),
-      denominator: Long.fromInt(3),
-    },
-    unbondingPeriod: {
-      seconds: new Long(unbondingPeriodSec),
-    },
-    trustingPeriod: {
-      seconds: new Long(trustPeriodSec),
-    },
-    maxClockDrift: {
-      seconds: new Long(20),
-    },
-    latestHeight: {
-      revisionNumber: new Long(0), // ??
-      revisionHeight: new Long(height),
-    },
-    proofSpecs: [iavlSpec, tendermintSpec],
-    upgradePath: ['upgrade', 'upgradedIBCState'],
-    allowUpdateAfterExpiry: false,
-    allowUpdateAfterMisbehaviour: false,
-  });
 }
 
 export async function prepareConnectionHandshake(
@@ -1095,56 +977,4 @@ export async function prepareChannelHandshake(
   // get a proof (for the proven height)
   const proof = await src.getChannelProof({ portId, channelId }, headerHeight);
   return proof;
-}
-
-interface ParsedAttribute {
-  readonly key: string;
-  readonly value: string;
-}
-
-export interface ParsedEvent {
-  readonly type: string;
-  readonly attributes: readonly ParsedAttribute[];
-}
-
-export function parsePacket({ type, attributes }: ParsedEvent): Packet {
-  if (type !== 'send_packet') {
-    throw new Error(`Cannot parse event of type ${type}`);
-  }
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const attributesObj: any = attributes.reduce(
-    (acc, { key, value }) => ({
-      ...acc,
-      [key]: value,
-    }),
-    {}
-  );
-  const [timeoutRevisionNumber, timeoutRevisionHeight] =
-    attributesObj.packet_timeout_height?.split('-') ?? [];
-  return {
-    sequence: Long.fromNumber(attributesObj.packet_sequence ?? 0, true),
-    /** identifies the port on the sending chain. */
-    sourcePort: attributesObj.packet_src_port ?? '',
-    /** identifies the channel end on the sending chain. */
-    sourceChannel: attributesObj.packet_src_channel ?? '',
-    /** identifies the port on the receiving chain. */
-    destinationPort: attributesObj.packet_dst_port ?? '',
-    /** identifies the channel end on the receiving chain. */
-    destinationChannel: attributesObj.packet_dst_channel ?? '',
-    /** actual opaque bytes transferred directly to the application module */
-    data: toUtf8(attributesObj.packet_data ?? ''),
-    /** block height after which the packet times out */
-    timeoutHeight:
-      timeoutRevisionNumber && timeoutRevisionHeight
-        ? Height.fromPartial({
-            revisionNumber: timeoutRevisionNumber,
-            revisionHeight: timeoutRevisionHeight,
-          })
-        : undefined,
-    /** block timestamp (in nanoseconds) after which the packet times out */
-    timeoutTimestamp: Long.fromString(
-      attributesObj.packet_timeout_timestamp ?? '0',
-      true
-    ),
-  };
 }
